pyproject.toml
[tool.poetry]
name = "guitar-sound-ver2"
version = "0.1.0"
description = ""
authors = ["rokufv <rokufv@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.37.0"
langchain = "^0.2.11"
langchain-google-genai = "^1.0.8"
python-dotenv = "^1.0.1"
faiss-cpu = "^1.8.0"
beautifulsoup4 = "^4.12.3"
yt-dlp = "^2024.7.25"
demucs = "^4.0.1"
soundfile = "^0.12.1"
librosa = "^0.10.2"
matplotlib = "^3.9.1"
streamlit-audiorec = "^0.1.3"
dtw-python = "^1.4.1"
rank-bm25 = "^0.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
